Trait agilulf::AsyncDatabase

pub trait AsyncDatabase: Send + Sync {
    fn get(
        &self, 
        key: Slice
    ) -> Pin<Box<dyn Future<Output = Result<Slice>> + Send>>;

    fn put(
        &self, 
        key: Slice, 
        value: Slice
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send>>;

    fn scan(
        &self, 
        start: Slice, 
        end: Slice
    ) -> Pin<Box<dyn Future<Output = Vec<(Slice, Slice)>> + Send>>;

    fn delete(
        &self, 
        key: Slice
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send>>;
}

Abstraction layer for a AsyncDatabase. Every method return a Future.

The return type of these function are fixed as Pin<Box<dyn Future<Output = Result<_>> + Send + '_>> rather than a generic type for convenience.

Required methods

fn get(&self, key: Slice) -> Pin<Box<dyn Future<Output = Result<Slice>> + Send>>

fn put(
&self,
key: Slice,
value: Slice
) -> Pin<Box<dyn Future<Output = Result<()>> + Send>>

fn scan(
&self,
start: Slice,
end: Slice
) -> Pin<Box<dyn Future<Output = Vec<(Slice, Slice)>> + Send>>

SCAN don't return a Result because if nothing is found, an empty vector will be returned.

fn delete(&self, key: Slice) -> Pin<Box<dyn Future<Output = Result<()>> + Send>>

Loading content...